Allow the NIT-other network to be specified
When ABM reads an NIT-other table it uses the first one it gets in the
transport stream. If multiple NIT-other tables are being published,
which one gets used is effectively random. This allows a network to be
specified explicitly by adding an optional nit_other_network_id
attribute to the <provider> element.

This will work fine as long as only one NIT-other network is needed. The
current code design assumes a single other network, so if more than one
is ever required some redesign work will be necessary.